Rename UseUtilityClassRule to InstantiableUtilityClassRule
