+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/design/InstantiableUtilityClassTest.java
@@ -6,6 +6,6 @@ package net.sourceforge.pmd.lang.java.rule.design;
 
 import net.sourceforge.pmd.test.PmdRuleTst;
 
-class UseUtilityClassTest extends PmdRuleTst {
+class InstantiableUtilityClassTest extends PmdRuleTst {
     // no additional unit tests
 }
