+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/rule/design/xml/InstantiableUtilityClass.xml
